Minor refactor of code
