+++ main.go
@@ -2,14 +2,13 @@ package main
 
 import (
 	"github.com/nicholashoule/go-ginrest/build"
-	"github.com/nicholashoule/go-ginrest/routes"
 )
 
 // main ...
 func main() {
 	// Build information
-	build.Run()
+	build.Info()
 
 	// Call routes.Run
-	routes.Run()
+	Run()
 }
